+++ ownCloud/AppDelegate.swift
@@ -49,7 +49,7 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 
 		ThemeStyle.registerDefaultStyles()
 
-		if VendorServices.shared.isBranded, VendorServices.shared.hasBrandedLogin {
+		if VendorServices.shared.isBranded {
 			staticLoginViewController = StaticLoginViewController(with: StaticLoginBundle.defaultBundle)
 			navigationController = ThemeNavigationController(rootViewController: staticLoginViewController!)
 			navigationController?.setNavigationBarHidden(true, animated: false)
@@ -61,8 +61,11 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
 			rootViewController = navigationController
 		}
 
-		window?.rootViewController = rootViewController!
-		window?.makeKeyAndVisible()
+		if !UIDevice.current.isIpad {
+			// Only set up window on non-iPad devices
+			window?.rootViewController = rootViewController!
+			window?.makeKeyAndVisible()
+		}
 
 		ImportFilesController.removeImportDirectory()