+++ ownCloud/Client/ClientRootViewController.swift
@@ -25,7 +25,7 @@ protocol ClientRootViewControllerAuthenticationDelegate : class {
 	func handleAuthError(for clientViewController: ClientRootViewController, error: NSError, editBookmark: OCBookmark?, preferredAuthenticationMethods: [OCAuthenticationMethodIdentifier]?)
 }
 
-class ClientRootViewController: UITabBarController, BookmarkContainer, ToolAndTabBarToggling {
+class ClientRootViewController: UITabBarController, BookmarkContainer, ToolAndTabBarToggling, UINavigationControllerDelegate {
 
 	// MARK: - Constants
 	let folderButtonsSize: CGSize = CGSize(width: 25.0, height: 25.0)
@@ -277,6 +277,18 @@ class ClientRootViewController: UITabBarController, BookmarkContainer, ToolAndTa
 		})
 	}
 
+	func navigationController(_ navigationController: UINavigationController,
+							  willShow viewController: UIViewController,
+							  animated: Bool) {
+		if viewController == emptyViewController {
+				closeClient()
+				if #available(iOS 13.0, *) {
+					// Prevent re-opening of items on next launch in case user has returned to the bookmark list
+					view.window?.windowScene?.userActivity = nil
+				}
+		}
+	}
+
 	func coreReady(_ lastVisibleItemId: String?) {
 		OnMainThread {
 			if let core = self.core {
@@ -290,7 +302,8 @@ class ClientRootViewController: UITabBarController, BookmarkContainer, ToolAndTa
 				}
 
 				let emptyViewController = self.emptyViewController
-				if VendorServices.shared.isBranded {
+				emptyViewController.navigationController?.delegate = self
+				if VendorServices.shared.isBranded, !VendorServices.shared.canAddAccount {
 					emptyViewController.navigationItem.title = "Manage".localized
 				} else {
 					emptyViewController.navigationItem.title = "Accounts".localized
@@ -442,6 +455,7 @@ extension ClientRootViewController : OCCoreDelegate {
 		if nsError == nil, let issueNSError = issue?.error as NSError? {
 			// Turn issues that are just converted authorization errors back into errors and discard the issue
 			if issueNSError.isOCError(withCode: .authorizationFailed) ||
+			   issueNSError.isOCError(withCode: .authorizationMethodNotAllowed) ||
 			   issueNSError.isOCError(withCode: .authorizationNoMethodData) ||
 			   issueNSError.isOCError(withCode: .authorizationMissingData) {
 				nsError = issueNSError
@@ -476,6 +490,12 @@ extension ClientRootViewController : OCCoreDelegate {
 
 			if nsError.isOCError(withCode: .authorizationNoMethodData) || nsError.isOCError(withCode: .authorizationMissingData) {
 				authFailureMessage = "No authentication data has been found for this connection.".localized
+
+				isAuthFailure = true
+			}
+
+			if nsError.isOCError(withCode: .authorizationMethodNotAllowed) {
+				authFailureMessage = NSString(format: "Authentication with %@ is no longer allowed. Re-authentication needed.".localized as NSString, core.connection.authenticationMethod?.name ?? "??") as String
 
 				isAuthFailure = true
 			}