+++ ownCloud/Settings/LogSettingsViewController.swift
@@ -64,7 +64,7 @@ class LogSettingsViewController: StaticTableViewController {
 		}, title: "Enable logging".localized, value: LogSettingsViewController.loggingEnabled, identifier: "enable-logging"))
 
 		let footerTitle = "When activated, logs may impact performance and include sensitive information. However the logs are not subject to automatic submission to %@ servers. Sharing logs with others is sole user responsibility.".localized
-		loggingSection.footerTitle = footerTitle.replacingOccurrences(of: "%@", with: OCAppIdentity.shared.appName ?? "ownCloud")
+		loggingSection.footerTitle = footerTitle.replacingOccurrences(of: "%@", with: VendorServices.shared.appName)
 
 		// Update section visibility
 		self.addSection(loggingSection)